feat: implement proper authentication for API endpoints

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -18,13 +19,22 @@ review_model_update = api.model('ReviewUpdate', {
 
 @api.route('/')
 class ReviewList(Resource):
+    @jwt_required()
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
     def post(self):
         """Register a new review"""
+        current_user = get_jwt_identity()
         try:
+            api.payload["user_id"] = current_user
             review = facade.create_review(api.payload)
+            place = facade.get_place(review.place_id)
+            if place.owner.id == current_user:
+                return {"Error": 'You cannot review your own place'}, 400
+            already_reviewed = any(review.get('user_id') == current_user for review in place.reviews)
+            if already_reviewed:
+                return {"Error": "You have already reviewed this place"}, 400
             return review.to_dict(), 201
         except (ValueError, TypeError) as e:
             return {"Error": str(e)}, 400
@@ -46,13 +56,18 @@ class ReviewResource(Resource):
             return {"Error": "Review not found"}, 404
         return review.to_dict(), 200
 
+    @jwt_required()
     @api.expect(review_model_update)
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
         try:
+            current_review = facade.get_review(review_id)
+            if current_review.user_id != current_user:
+                return {"Error": 'Unauthorized action'}, 403
             review = facade.update_review(review_id, api.payload)
             return review.to_dict(), 200
         except ValueError as e:
@@ -60,11 +75,16 @@ class ReviewResource(Resource):
         except TypeError as e:
             return {"Error": str(e)}, 404
 
+    @jwt_required()
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
         try:
+            review = facade.get_review(review_id)
+            if review.user_id != current_user:
+                return {"Error": 'Unauthorized action'}, 403
             facade.delete_review(review_id)
             return {"Message": "Review deleted successfully"}, 200
         except TypeError:
